requirements.txt
PyQt5>=5.15.0
opencv-python>=4.5.0
qrcode>=7.3.0
pandas>=1.5.0
matplotlib>=3.5.0
numpy>=1.20.0
reportlab>=3.6.0
pillow>=9.0.0
pyinstaller>=5.0.0
pyzbar>=0.1.9  # Optional, for improved QR code scanning